feat(profiler_builtin): support nanosecond accuracy

+++ src/misc/lv_profiler_builtin_private.h
@@ -32,7 +32,7 @@ extern "C" {
 struct _lv_profiler_builtin_config_t {
     size_t buf_size;                    /**< The size of the buffer used for profiling data */
     uint32_t tick_per_sec;              /**< The number of ticks per second */
-    uint32_t (*tick_get_cb)(void);      /**< Callback function to get the current tick count */
+    uint64_t (*tick_get_cb)(void);      /**< Callback function to get the current tick count */
     void (*flush_cb)(const char * buf); /**< Callback function to flush the profiling data */
     int (*tid_get_cb)(void);            /**< Callback function to get the current thread ID */
     int (*cpu_get_cb)(void);            /**< Callback function to get the current CPU */
